Pull out behaviour into separate file for easier testing
